Split networking module into smaller submodules.
