+++ src/hotstuff/messages.rs
@@ -10,7 +10,10 @@ use std::mem;
 use borsh::{BorshDeserialize, BorshSerialize};
 
 use crate::{
-    networking::{Cacheable, ProgressMessage},
+    networking::{
+        messages::ProgressMessage,
+        receiving::Cacheable,
+    },
     types::{basic::*, block::*, keypair::*, signed_messages::SignedMessage},
 };
