+++ src/networking/mod.rs
@@ -1,0 +1,7 @@
+pub mod network;
+
+pub mod messages;
+
+pub(crate) mod receiving;
+
+pub(crate) mod sending;
