+++ tests/common/network.rs
@@ -10,7 +10,10 @@ use std::{
 
 use ed25519_dalek::VerifyingKey;
 use hotstuff_rs::{
-    networking::{Message, Network},
+    networking::{
+        messages::Message,
+        network::Network,
+    },
     types::validators::{ValidatorSet, ValidatorSetUpdates},
 };
